Updated Utility.java and MenuManager.java
Utility now has a capitalize(String) method for capitalizing the first letter of a given String argument.

Renamed closeAllMenus to closeMenus() in MenuManager. The MenuManager might get changed in the future to function with a single HashSet or HashMap for storing AbstractMenus and more easily accessing or iterating over them.